kernel: net: make in_aton() 32-bit internally
[ Upstream e013fb7c4c ]

+++ linux/linux-2.6.22.19/net/core/utils.c
@@ -50,7 +50,7 @@ EXPORT_SYMBOL(net_ratelimit);
 
 __be32 in_aton(const char *str)
 {
-	unsigned long l;
+	unsigned int l;
 	unsigned int val;
 	int i;
